insights/insights, standalone/standalone - rename loaders
