+++ src/entry-standalone.tsx
@@ -3,7 +3,7 @@ import * as ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { I18nProvider } from '@lingui/react';
 import { i18n } from '@lingui/core';
-import App from './loaders/standalone/standalone-loader';
+import App from './loaders/standalone/loader';
 import 'src/l10n';
 
 // Entrypoint for compiling the app to run in standalone mode
