+++ src/loaders/insights/loader.tsx
