+++ src/loaders/standalone/loader.tsx
